+++ fractal_server/tasks/v2/ssh/deactivate_pixi.py
@@ -65,6 +65,7 @@ def deactivate_ssh_pixi(
             with SingleUseFractalSSH(
                 ssh_config=SSHConfig(
                     host=resource.host,
+                    port=resource.port_or_default,
                     user=profile.username,
                     key_path=profile.ssh_key_path,
                 ),