+++ fractal_server/tasks/v2/ssh/delete.py
@@ -60,6 +60,7 @@ def delete_ssh(
             with SingleUseFractalSSH(
                 ssh_config=SSHConfig(
                     host=resource.host,
+                    port=resource.port_or_default,
                     user=profile.username,
                     key_path=profile.ssh_key_path,
                 ),