+++ fractal_server/tasks/v2/ssh/reactivate_pixi.py
@@ -74,6 +74,7 @@ def reactivate_ssh_pixi(
             with SingleUseFractalSSH(
                 ssh_config=SSHConfig(
                     host=resource.host,
+                    port=resource.port_or_default,
                     user=profile.username,
                     key_path=profile.ssh_key_path,
                 ),